backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai>=1.0.0
spotipy==2.25.1
lyricsgenius==3.6.4
pandas==2.3.0
pinecone
numpy==2.3.1
scikit-learn==1.7.0
scipy==1.16.0
regex==2024.11.6
tqdm==4.67.1
joblib==1.5.1
requests==2.32.4
matplotlib==3.10.3



# Database + ORM
psycopg2-binary
sqlalchemy